pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "epigraph"
version = "0.1.0"
description = "Hybrid TypeDB + columnar Parquet toolkit for DNA methylation biomarker discovery"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "The epigraph authors" }]
dependencies = [
    # Core data processing
    "polars>=1.0",
    "pyarrow>=15.0",
    "duckdb>=1.0",
    "pandas>=2.0",            # for openpyxl compatibility
    "openpyxl>=3.1",          # XLSX reading
    "numpy>=1.26",

    # Statistics
    "scipy>=1.12",
    "statsmodels>=0.14",

    # AWS / S3 (optional at runtime — lazy-imported in common/io.py)
    "boto3>=1.34",

    # Configuration
    "pyyaml>=6.0",
    "pydantic>=2.6",

    # Visualisation
    "matplotlib>=3.8",
    "seaborn>=0.13",

    # CLI
    "click>=8.1",
    "rich>=13.0",

    # Logging
    "structlog>=24.1",
]

[project.optional-dependencies]
typedb = [
    "typedb-driver>=3.0",
]
report = [
    "python-docx>=1.1",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.10",
    "mypy>=1.8",
    # Type stubs (silence ~40% of mypy's import-untyped noise).
    "types-PyYAML>=6.0",
    "types-openpyxl>=3.1",
    "pandas-stubs>=2.0",
    "ipykernel>=6.29",
]
notebooks = [
    "jupyter>=1.0",
    "plotly>=5.18",
]
all = [
    "epigraph[typedb,report,dev,notebooks]",
]

[project.scripts]
create-dev-subset = "epigraph.db_build.create_dev_subset:main"
ingest-clinical = "epigraph.db_build.load_clinical_metadata:main"
build-annotations = "epigraph.db_build.parse_annotations:main"
import-typedb = "epigraph.db_build.import_typedb:main"
convert-beta = "epigraph.db_build.convert_beta_to_parquet:main"
dataset-stats = "epigraph.db_build.dataset_stats:main"
filter-cpgs = "epigraph.db_build.filter_cpgs:main"
map-cpg-genes = "epigraph.db_build.map_cpg_to_genes:main"
derive-features = "epigraph.analysis.feature_aggregation:main"
run-analysis = "epigraph.analysis.cohort_comparison:main"
run-pipeline = "epigraph.pipeline:main"
pathway-enrichment = "epigraph.analysis.pathway_enrichment:main"
biomarker-candidates = "epigraph.analysis.biomarker_candidates:main"
hypermethylation = "epigraph.analysis.hypermethylation:main"
generate-report = "epigraph.analysis.generate_report:main"
visualise = "epigraph.analysis.visualise:main"

[tool.hatch.build.targets.wheel]
packages = ["src/epigraph"]

[tool.hatch.build.targets.sdist]
include = ["src/epigraph"]

[tool.hatch.build]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = ["-p", "no:asyncio"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
